:recycle: Refactorización inicial
Detalles adicionales:
♻ Se refactoriza la capa model.
